requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
packaging==24.1
phonenumbers==8.13.45
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
whitenoise==6.7.0
